agent/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
apscheduler>=3.11.2
firebase-admin>=7.2.0
python-dotenv>=1.2.1
requests>=2.31.0
langchain>=1.2.10
google-generativeai>=0.8.6
pydantic>=2.10.0
typing-extensions>=4.12.0
groq>=0.11.0
